Refactor Comparators
* group comparators within EntityComparators class
