+++ wegas-core/src/main/java/com/wegas/reviewing/persistence/Review.java
@@ -11,7 +11,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.wegas.core.exception.client.WegasIncompatibleType;
 import com.wegas.core.persistence.AbstractEntity;
 import com.wegas.core.persistence.DatedEntity;
-import com.wegas.core.persistence.EntityCreatedTimeComparator;
+import com.wegas.core.persistence.EntityComparators;
 import com.wegas.core.persistence.ListUtils;
 import com.wegas.reviewing.persistence.evaluation.EvaluationInstance;
 
@@ -173,7 +173,7 @@ public class Review extends AbstractEntity implements DatedEntity {
      * @return the list of evaluation instance composing the feedback
      */
     public List<EvaluationInstance> getFeedback() {
-        Collections.sort(this.feedback, new EntityCreatedTimeComparator<>());
+        Collections.sort(this.feedback, new EntityComparators.CreateTimeComparator<>());
         return this.feedback;
     }
 
@@ -192,7 +192,7 @@ public class Review extends AbstractEntity implements DatedEntity {
      * @return the list of evaluation instances composing the feedback comments
      */
     public List<EvaluationInstance> getComments() {
-        Collections.sort(this.comments, new EntityCreatedTimeComparator<>());
+        Collections.sort(this.comments, new EntityComparators.CreateTimeComparator<>());
         return this.comments;
     }
